Modified - move convertor of number to core
