+++ lib/core/screens/splash_Screen.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 
 import '../theme/app_theme.dart';
-import '../../widgets/en_to_ar_number_convertor.dart';
+import '../logic/en_to_ar_number_convertor.dart';
 import '../../widgets/splashscreen.dart';
 import 'navigation_bottom_screen.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
