+++ lib/features/customer_feature/presentation/screens/customer_detail_order_screen.dart
@@ -6,7 +6,7 @@ import '../../../../models/customer.dart';
 import '../../../../models/order.dart';
 import '../../../../core/theme/app_theme.dart';
 import '../providers/customer_info.dart';
-import '../../../../widgets/en_to_ar_number_convertor.dart';
+import '../../../../core/logic/en_to_ar_number_convertor.dart';
 import '../../../../screens/order_view_screen.dart';
 
 class CustomerDetailOrderScreen extends StatefulWidget {
