+++ lib/screens/cart_screen.dart
@@ -14,7 +14,7 @@ import '../screens/order_products_send_screen.dart';
 import '../widgets/card_item.dart';
 import '../widgets/custom_dialog_enter.dart';
 import '../widgets/custom_dialog_profile.dart';
-import '../widgets/en_to_ar_number_convertor.dart';
+import '../core/logic/en_to_ar_number_convertor.dart';
 import '../core/widgets/main_drawer.dart';
 
 class CartScreen extends StatefulWidget {
