+++ lib/screens/charity_screen.dart
@@ -8,7 +8,7 @@ import 'package:tamizshahr/widgets/charity_item_Charities_screen.dart';
 import '../models/search_detail.dart';
 import '../provider/Products.dart';
 import '../core/theme/app_theme.dart';
-import '../widgets/en_to_ar_number_convertor.dart';
+import '../core/logic/en_to_ar_number_convertor.dart';
 import '../core/widgets/main_drawer.dart';
 
 class CharityScreen extends StatefulWidget {
