+++ lib/screens/clear_screen.dart
@@ -16,7 +16,7 @@ import 'package:tamizshahr/widgets/custom_dialog_send_request.dart';
 
 import '../models/search_detail.dart';
 import '../core/theme/app_theme.dart';
-import '../widgets/en_to_ar_number_convertor.dart';
+import '../core/logic/en_to_ar_number_convertor.dart';
 import '../core/widgets/main_drawer.dart';
 import '../features/customer_feature/presentation/screens/login_screen.dart';
 import '../core/screens/navigation_bottom_screen.dart';
