+++ lib/screens/collect_detail_screen.dart
@@ -5,7 +5,7 @@ import 'package:provider/provider.dart';
 import 'package:tamizshahr/features/waste_feature/business/entities/request_waste_item.dart';
 import 'package:tamizshahr/provider/wastes.dart';
 import 'package:tamizshahr/widgets/collect_details_collects_item.dart';
-import 'package:tamizshahr/widgets/en_to_ar_number_convertor.dart';
+import 'package:tamizshahr/core/logic/en_to_ar_number_convertor.dart';
 
 import '../core/theme/app_theme.dart';
 import '../core/widgets/main_drawer.dart';
