+++ lib/screens/product_detail_screen.dart
@@ -10,7 +10,7 @@ import 'package:tamizshahr/widgets/buton_bottom.dart';
 import '../models/product.dart';
 import '../provider/Products.dart';
 import '../core/theme/app_theme.dart';
-import '../widgets/en_to_ar_number_convertor.dart';
+import '../core/logic/en_to_ar_number_convertor.dart';
 import '../core/widgets/main_drawer.dart';
 import 'cart_screen.dart';
