+++ lib/screens/waste_cart_screen.dart
@@ -12,7 +12,7 @@ import '../provider/wastes.dart';
 import '../features/waste_feature/presentation/wastes_screen.dart';
 import '../widgets/custom_dialog_enter.dart';
 import '../widgets/custom_dialog_profile.dart';
-import '../widgets/en_to_ar_number_convertor.dart';
+import '../core/logic/en_to_ar_number_convertor.dart';
 import '../core/widgets/main_drawer.dart';
 import '../widgets/waste_cart_item.dart';
 import 'address_screen.dart';
