+++ lib/widgets/article_item_article_screen.dart
@@ -6,7 +6,7 @@ import 'package:shamsi_date/shamsi_date.dart';
 import '../models/article/article.dart';
 import '../core/theme/app_theme.dart';
 import '../screens/article_detail_screen.dart';
-import 'en_to_ar_number_convertor.dart';
+import '../core/logic/en_to_ar_number_convertor.dart';
 
 class ArticleItemArticlesScreen extends StatelessWidget {
   @override
