+++ lib/widgets/custom_dialog_pay_charity.dart
@@ -3,7 +3,7 @@ import 'package:intl/intl.dart' as intl;
 import 'package:tamizshahr/models/charity.dart';
 
 import '../core/theme/app_theme.dart';
-import 'en_to_ar_number_convertor.dart';
+import '../core/logic/en_to_ar_number_convertor.dart';
 
 class CustomDialogPayCharity extends StatefulWidget {
   final int totalWallet;
