+++ lib/widgets/message_item.dart
@@ -3,7 +3,7 @@ import 'package:shamsi_date/shamsi_date.dart';
 
 import '../models/message.dart';
 import '../core/theme/app_theme.dart';
-import '../widgets/en_to_ar_number_convertor.dart';
+import '../core/logic/en_to_ar_number_convertor.dart';
 
 class MessageItem extends StatelessWidget {
   const MessageItem({
