+++ lib/widgets/transaction_item_transactions_screen.dart
@@ -4,7 +4,7 @@ import 'package:provider/provider.dart';
 import 'package:tamizshahr/models/transaction.dart';
 import 'package:tamizshahr/core/theme/app_theme.dart';
 
-import 'en_to_ar_number_convertor.dart';
+import '../core/logic/en_to_ar_number_convertor.dart';
 
 class TransactionItemTransactionsScreen extends StatelessWidget {
   @override
