+++ lib/widgets/waste_cart_item_animated_list.dart
@@ -7,7 +7,7 @@ import 'package:tamizshahr/features/waste_feature/business/entities/wasteCart.da
 
 import '../core/theme/app_theme.dart';
 import '../provider/wastes.dart';
-import 'en_to_ar_number_convertor.dart';
+import '../core/logic/en_to_ar_number_convertor.dart';
 
 class WasteCartItemAnimatedList extends StatefulWidget {
   final WasteCart wasteItem;
